+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `cc1101-transmit` command:
+  - configure transceiver after reading from stdin
+    to avoid delay between configuration and transmission, if pipe is slow
 
 ## [2.5.0] - 2020-12-30
 ### Added